+++ courseSelection/database/migrations/2023_05_18_025559_student_course_table.php
@@ -12,8 +12,8 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('student_courses', function (Blueprint $table) {
-            $table->integer('studentId');
-            $table->integer('courseId');
+            $table->bigInteger('studentId');
+            $table->bigInteger('courseId');
             $table->timestamps();
         });
     }
